Reformatting and documentation
